[DF] Retrieve RDatasetSpec for RDF built from JSON
