+++ tree/dataframe/inc/ROOT/RDF/Utils.hxx
@@ -31,7 +31,9 @@
 class TTree;
 class TTreeReader;
 
-
+namespace ROOT::RDF::Experimental {
+class RDatasetSpec;
+}
 namespace ROOT {
 namespace RDF {
 using ColumnNames_t = std::vector<std::string>;
@@ -324,6 +326,16 @@ auto MakeAliasedSharedPtr(T *rawPtr)
    const static std::shared_ptr<T> fgRawPtrCtrlBlock;
    return std::shared_ptr<T>(fgRawPtrCtrlBlock, rawPtr);
 }
+
+
+/**
+ * \brief Function to retrieve RDatasetSpec from JSON file provided
+ * \param[in] jsonFile Path to the dataset specification JSON file.
+ *
+ * This function allows us to have access to an RDatasetSpec which needs to
+ * be created when we use the FromSpec factory function.
+ */
+ROOT::RDF::Experimental::RDatasetSpec RetrieveSpecFromJson(const std::string &jsonFile);
 
 } // end NS RDF
 } // end NS Internal
